re-structure code
