+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "rust-analyzer.linkedProjects": [
         ".\\Cargo.toml",
+        ".\\Cargo.toml",
         ".\\Cargo.toml"
     ]
 }
